+++ tests/test_file.c
@@ -23,27 +23,12 @@ void test_file_open()
   file_close(file);
 }
 
-void test_file_replace_buffer()
-{
-  file_t* file = NULL;
-  TEST_ASSERT(file_open(&file, "filethatexist.log") == true);
-
-  data_t buff = { 64, NULL};
-  TEST_ASSERT(data_alloc(&buff) == true);
-  memset(buff.address, 72, 64);
-  TEST_CHECK(file_replace_buffer(file, &buff) == true);
-
-  data_free(&buff);
-  file_close(file);
-}
-
 void test_file_write()
 {
   file_t* file = NULL;
   TEST_ASSERT(file_open(&file, "filethatexist.log") == true);
 
-  file->buffer.size += 64;
-  TEST_ASSERT(data_alloc(&file->buffer) == true);
+  TEST_ASSERT(data_alloc(&file->buffer, file->buffer.size + 64) == true);
   memset(file->buffer.address, 72, 64);
 
   TEST_CHECK(file_write(file, 0, 32, false) == true);
@@ -74,14 +59,13 @@ void test_file_read_line()
   file_t* file = NULL;
   TEST_ASSERT(file_open(&file, "filethatexist.log") == true);
 
-  file->buffer.size = 1;
-  TEST_ASSERT(data_alloc(&file->buffer) == true);
+  TEST_ASSERT(data_alloc(&file->buffer, 1) == true);
   memset(file->buffer.address, 10, 1);
   TEST_CHECK(file_write(file, 4, 1, false) == true);
 
   TEST_CHECK(file_read_line(file, true) == true);
+  TEST_CHECK(file->buffer.size == 5);
   TEST_CHECK(file->offset == 5);
-  TEST_CHECK(file->buffer.size = 5);
 
   file_close(file);
 }
@@ -89,7 +73,6 @@ void test_file_read_line()
 TEST_LIST = {
   { "file_create", test_file_create },
   { "file_open", test_file_open },
-  { "file_replace_buffer", test_file_replace_buffer },
   { "file_write", test_file_write },
   { "file_read", test_file_read },
   { "file_read_line", test_file_read_line },